requirements.txt
convert-numbers==0.4
Pillow==8.4.0
tk==0.1.0
